\name{BcnDataAccess-package}
\alias{BcnDataAccess-package}
\alias{BcnDataAccess}
\docType{package}
\title{
Easing access to Bcn data
}
\description{
This packages paves the way to access any data related to Barcelona. It is part of the BCNAnalytics community.
}
\details{
\tabular{ll}{
Package: \tab DataAccess\cr
Type: \tab Package\cr
Version: \tab 0.8\cr
Date: \tab 2014-10-19\cr
}
This package provides a tree to get data from different sources in a straight way, so that people from the BCNAnalytics (www.bcnanalytics.com) community can share data bases or code with wuch data bases.
}
\author{
Maintainer: Aleix Ruiz de Villa <aleixrvr@gmail.com>
}
\references{
http://bcnanalytics.com/
}
\examples{
BcnDataSources$Ajuntament$AturAgost
BcnDataSources$Ajuntament$AturAgost$getData()
}
